requirements.txt
gitpython
rich
google-generativeai
